refactor: pull out various convert steps into helper functions
This will be useful soon for a modernize-js command that uses some of those
steps.
